Debugging

+++ resources/views/categories/index.blade.php
@@ -11,6 +11,7 @@
                   <div id="search">
                     <p>Search</p>
                     <input type="text" class="mb-10 outline-0 px-2" placeholder="Search for products" style="border: 1px solid gray" name="q" id="q">
+                      <a href="/admin/category/create" class="create">Create!</a>
                   </div>
                   <table id="example1" class="table table-bordered table-striped">
                     <thead>
@@ -59,8 +60,11 @@
       
       <script>
         function deleteItem(url) {
-          const request = new axiosWrapper(url);
-          request.delete(5);
+          let isConfirmed = confirm("Do you really want to delete the category permanently? All products with this category will also be deleted!");
+          if(isConfirmed) {
+            const request = new axiosWrapper(url);
+            request.delete(5);
+          }
         }
       </script>
       <script>
